HIVE-28724:CREATE TABLE is not handled by DDLSemanticAnalyzerFactory
